Corrige error al salir de callejero sin usuario registrado

+++ frontend/src/routes/explore/[code]/street/[[name]]/+page.svelte
@@ -59,8 +59,10 @@
   })
 
   beforeNavigate(() => {
-    const url = `/papi/street/${$street.mun_code}/${$street.cat_name}/lock`
-		fetch(url, {method: 'DELETE'})
+    if (data.user) {
+      const url = `/papi/street/${$street.mun_code}/${$street.cat_name}/lock`
+      fetch(url, {method: 'DELETE'})
+    }
   })
